requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.1
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pip==21.2.4
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
setuptools==58.0.4
sqlparse==0.5.3
typing_extensions==4.12.2